Prepare laserlogbook files for pollyxt_tjk
This commit was used to add the feature to prepare the laserlogbook file for pollyxt_tjk. Due to some reason, the laserlogbook files were not generated for this system. In instead, it produces the `yyyymmdd_temps.txt` file for saving the housekeeping data. And the format is also different than the laserlogbook book. Therefore, in order to still showing the housekeeping plot for this lidar, this update was implemented.


Former-commit-id: 2d00ca02dec6fe4f46d022d942199c6c14f97223
Former-commit-id: cb9eab587106d1d2f1ebd2f2c7c79e0f162037ce
Former-commit-id: bd18d29616070fc492463974e806222bde05d92e