Prepare laserlogbook files for pollyxt_tjk
This commit was used to add the feature to prepare the laserlogbook file for pollyxt_tjk. Due to some reason, the laserlogbook files were not generated for this system. In instead, it produces the `yyyymmdd_temps.txt` file for saving the housekeeping data. And the format is also different than the laserlogbook book. Therefore, in order to still showing the housekeeping plot for this lidar, this update was implemented.