Update: fix timestamp
Enable to fix the timestamp for each profile, instead of using the
original timestamp acquired from PC. The original timestamp suffers from
the instability of the PC hardware and often has bias of several
seconds.

Meanwhile, a new check of MShots has been implemented, which will
average the profiles if the MShots is less than 600.


Former-commit-id: d7380bd6df5720f6dd7781d619cb6e8fe5dc994f
Former-commit-id: f66e59e55da032b589bc8e7dfeedcf81c34e0af4
Former-commit-id: 8c66b3de30ef75ef8ed444a8bbc8ebdcc651900e