Update: fix timestamp
Enable to fix the timestamp for each profile, instead of using the
original timestamp acquired from PC. The original timestamp suffers from
the instability of the PC hardware and often has bias of several
seconds.

Meanwhile, a new check of MShots has been implemented, which will
average the profiles if the MShots is less than 600.