Improvement: merge link and history file
In this commit, the `pollynet_history_and_places_new.txt` and
`pollynet_processing_chain_config_link.xlsx` were merged into
`pollynet_processing_chain_config_links.xlsx`. Old style was not
supported any more.

By this update, all campaign information should be registered in the new **link** file, in which the items are consistent with **history** file. But the duplicated entries, due to the variable V*, are removed.

Details can be found in the new **link** file.


Former-commit-id: f1a5c37d55471ed01d8c149072b8b21c57b6a37d
Former-commit-id: 0ae45f449addf99937a36c283c1a6d37ed512684
Former-commit-id: 06c649a1ac88e7f2581f278ff209232bbbab843e