Improvement: merge link and history file
In this commit, the `pollynet_history_and_places_new.txt` and
`pollynet_processing_chain_config_link.xlsx` were merged into
`pollynet_processing_chain_config_links.xlsx`. Old style was not
supported any more.

By this update, all campaign information should be registered in the new **link** file, in which the items are consistent with **history** file. But the duplicated entries, due to the variable V*, are removed.

Details can be found in the new **link** file.